# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Work Stat Lite Api Documentation",
        "description": "Work Stat Lite Api Documentation",
        "contact": {
            "email": "orutu1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://wslite-backend.herokuapp.com/api//",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/departments": {
            "get": {
                "tags": [
                    "Show All Departments"
                ],
                "description": "Show all departments",
                "operationId": "AllDepartments",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Sorry no record was found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Create department"
                ],
                "description": "Create  department",
                "operationId": "CreateDepartment",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter department name"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Enter department description"
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "Enter department  parent_id"
                    },
                    {
                        "name": "hod_id",
                        "in": "query",
                        "description": "Enter department hod_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Department created successfully"
                    },
                    "400": {
                        "description": "Sorry unable to create department"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/departments/{id}": {
            "get": {
                "tags": [
                    "Show Department"
                ],
                "description": "Show single department",
                "operationId": "ShowDepartment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter department id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Sorry this record do not exist"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Delete Department"
                ],
                "summary": "Delete existing department",
                "description": "Deletes a department and returns no content",
                "operationId": "deleteDepartment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Department Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Sorry this records do not exist"
                    },
                    "401": {
                        "description": "Sorry unable to delete department"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Update department"
                ],
                "description": "Update  department",
                "operationId": "UpdateDepartment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter id"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter department name"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Enter department description"
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "Enter department  parent_id"
                    },
                    {
                        "name": "hod_id",
                        "in": "query",
                        "description": "Enter department hod_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Department updated successfully"
                    },
                    "400": {
                        "description": "Sorry unable to update department"
                    },
                    "404": {
                        "description": "Sorry this data do not exist"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}